requirements.txt
openai>=1.0.0
chromadb==0.5.23
python-dotenv>=1.0.0
rank-bm25>=0.2.2
PyYAML>=6.0
sentence-transformers>=3.0.0
streamlit>=1.38.0
pypdf>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
openpyxl>=3.1.0
httpx>=0.27.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
olefile>=0.47

# Optional: local model / fine-tuning experiments.
# Install these only in a GPU environment if Scenario A FT is required.
transformers>=4.45.0
accelerate>=0.34.0
datasets>=2.20.0
peft>=0.12.0
bitsandbytes>=0.43.0
